feat: SCO-20 - initial spike and update to pf5
This commit updates the seed project to patternfly 5 as well as upgrades
several other dependencies on the existing seed project.  This change
also involved updating the code to work with react-router v6 and uses a
much simplified approach to app routing, a trend that should be carried
through to the AppLayout component.  The update to React 18
introduces a breaking change to the supplied tests.  The tests rely on
enzyme which is discontinued and will need to be migrated to something
else, perhaps react testing library or cypress.