fix: SCO-20 update storybook
This change updates storybook to the latest version and applies several
migration scripts that assist with this.  However the existing stories
need to be updated, and it would be better to move them to be closer to
the components they document.  With this commit storybook builds,
however the development mode does not work yet.